fixes #127 - Configure tool param handling
Configure tool will now surround parameters by double quotes to avoid
commas in password or email or devicename to break JSON format. With
this change, spaces before and after the semicolumn separator in
MyJDownloader configuration JSON file are forbidden (therefore, changed
in the dist default file).

This new behaviour, however, is buggy with double quotes in password.
This is a limitation of sed that does not allow non greedy regexp. This
means if a password contains a double quote, it will be correctly handle
the first time only and will break everything on password change. Same
for devicename or email.

Let's pray for JDownloader to NOT change back with the space on the
semicolumn separator or everything will be broken again.

I know this is not perfect but fixes 99% of issues to avoid
recompiling the universe or adding a thousand dependencies.